refactor: avoid a whole bunch of circular imports
