+++ src/containers/Masthead/drawer/drawerMenuTypes.ts
@@ -6,4 +6,12 @@
  *
  */
 
+import { GQLSubjectMenu_RootFragment } from "../../../graphqlTypes";
+
 export type MenuType = "programme" | "om" | "subject";
+
+export type AllTopicsType = NonNullable<GQLSubjectMenu_RootFragment["allTopics"]>[0];
+
+export type TopicWithSubTopics = AllTopicsType & {
+  subtopics: TopicWithSubTopics[];
+};
