+++ src/containers/MyNdla/Learningpath/components/LearningpathItem.tsx
@@ -6,17 +6,15 @@
  *
  */
 
-import { Fragment, useMemo } from "react";
+import { Fragment, ReactNode, useMemo } from "react";
 import { useTranslation } from "react-i18next";
 import { PencilLine, CheckLine, UserLine, RouteLine } from "@ndla/icons";
 import { ListItemContent, ListItemHeading, ListItemRoot, ListItemVariantProps, Text } from "@ndla/primitives";
 import { SafeLink } from "@ndla/safelink";
 import { styled } from "@ndla/styled-system/jsx";
 import { linkOverlay } from "@ndla/styled-system/patterns";
-import { useLearningpathActionHooks } from "./LearningpathActionHooks";
 import { GQLMyNdlaLearningpathFragment } from "../../../../graphqlTypes";
 import { routes } from "../../../../routeHelpers";
-import SettingsMenu from "../../components/SettingsMenu";
 import {
   LEARNINGPATH_PRIVATE,
   LEARNINGPATH_READY_FOR_SHARING,
@@ -55,11 +53,10 @@ const MenuWrapper = styled("div", {
 
 interface Props {
   learningpath: GQLMyNdlaLearningpathFragment;
-  showMenu: boolean;
+  menu?: ReactNode;
 }
-export const LearningpathItem = ({ learningpath, showMenu = true, context, ...rest }: Props & ListItemVariantProps) => {
+export const LearningpathItem = ({ learningpath, context, menu, ...rest }: Props & ListItemVariantProps) => {
   const { t, i18n } = useTranslation();
-  const menuItems = useLearningpathActionHooks(learningpath);
 
   const MaybeWrapper = context === "list" ? "li" : Fragment;
 
@@ -116,11 +113,7 @@ export const LearningpathItem = ({ learningpath, showMenu = true, context, ...re
             </StatusText>
           ) : null}
         </ListItemContent>
-        {showMenu ? (
-          <MenuWrapper>
-            <SettingsMenu menuItems={menuItems} />
-          </MenuWrapper>
-        ) : null}
+        {menu ? <MenuWrapper>{menu}</MenuWrapper> : null}
       </MaybeWrapper>
     </ListItemRoot>
   );
