+++ src/util/error/index.ts
@@ -10,4 +10,3 @@ import { ApolloError } from "@apollo/client";
 import { NDLAError } from "./NDLAError";
 
 export type ErrorType = ApolloError | Error | NDLAError | string | unknown;
-export type LogLevel = "error" | "warn" | "info";
